requirements.txt
fastapi
uvicorn
python-multipart
olefile
reportlab
Pillow
